+++ projects/services-owc-json/src/lib/owc-json.service.spec.ts
@@ -1,16 +1,22 @@
 import { TestBed, getTestBed } from '@angular/core/testing';
 
 import { OwcJsonService } from './owc-json.service';
-import { IOwsContext, IOwsResource, IOwsOffering } from '@ukis/datatypes-owc-json';
-import { barebonesContext, basicContext } from '../../assets/exampleContext';
+import { IOwsContext,  } from '@ukis/datatypes-owc-json';
+import { barebonesContext, basicContext, exampleContext } from '../../assets/exampleContext';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { optimizeGroupPlayer } from '@angular/animations/browser/src/render/shared';
+import { Fill, Stroke, Style } from 'ol/style.js';
+import { coastalXTestContext } from '../../assets/coastalx.test.context';
+import { iterateListLike } from '@angular/core/src/change_detection/change_detection_util';
+import { osm } from '@ukis/base-layers-raster';
+import { LayersService } from '@ukis/services-layers';
+import { VectorLayer, WfsLayertype, GeojsonLayertype } from '@ukis/datatypes-layers';
+import { Feature, Polygon, FeatureCollection } from 'geojson';
 
 
 
-
-describe('OwcJsonService', () => {
-  const allTestContexts = [barebonesContext, basicContext];
+describe('OwcJsonService: reading data from owc', () => {
+  const allTestContexts = [barebonesContext, basicContext, exampleContext, coastalXTestContext];
   let injector: TestBed;
   let service: OwcJsonService;
   
@@ -71,6 +77,7 @@ describe('OwcJsonService', () => {
 
   it('#createRasterLayerFromOffering should return an IRasterLayerOptions instance', () => {
     for(const context of allTestContexts) {
+      console.log("======= context: ", context.id)
       for(const resource of service.getResources(context)) {
         for(const offering of resource.properties.offerings) {
           if(service.checkIfServiceOffering(offering)) {
@@ -78,14 +85,23 @@ describe('OwcJsonService', () => {
             const operation = offering.operations[0];
             expect(operation).toBeTruthy();
             const rlayerOptions = service.createRasterLayerFromOffering(offering, resource);
-  
             expect(rlayerOptions.name).toBe(resource.properties.title);
             expect(rlayerOptions.id as string).toBe(resource.id as string);
             expect(rlayerOptions.opacity).toBe(1); // opacity is not encoded in owc-json per default. Allways falling back to 1. 
-            expect(rlayerOptions.visible).toBe(resource.properties.active);
+            console.log("======= check visible: ", resource.properties.active);
+            // if active is not set in conext file, then it must default to true
+            expect(rlayerOptions.visible).toBe(resource.properties.active === undefined ? true : resource.properties.active);
+
+            console.log("======= check removeable: ", rlayerOptions.removable);
             expect(rlayerOptions.removable).toBe(true); // "removable" is not encoded in owc-json; falling back to "true"
+            console.log("======= check filtertype == Overlays: ", rlayerOptions.filtertype);
             expect(rlayerOptions.filtertype).toBe("Overlays"); // all data in owc-json will - for now - be an overlay. Might be changed in the future. 
-            expect(rlayerOptions.type).toBe(offering.code.split("/").pop());
+
+            console.log("======= check serviceType: ", rlayerOptions.type);
+
+      
+            expect(rlayerOptions.type).toBe("wms");
+            console.log("======= check url: ", operation.href.substr(0, operation.href.indexOf("?")));
             expect(rlayerOptions.url).toBe(operation.href.substr(0, operation.href.indexOf("?")));
   
             // @TODO: checke hier auch noch die rlayerOptions.params
@@ -104,18 +120,23 @@ describe('OwcJsonService', () => {
 
             const operation = offering.operations[0];
             expect(operation).toBeTruthy();
-            const vlayerOptions = service.createVectorLayerFromOffering(offering, resource);
-  
-            expect(vlayerOptions.name).toBe(resource.properties.title);
-            expect(vlayerOptions.id as string).toBe(resource.id as string);
-            expect(vlayerOptions.opacity).toBe(1); // opacity is not encoded in owc-json per default. Allways falling back to 1. 
-            expect(vlayerOptions.visible).toBe(resource.properties.active);
-            expect(vlayerOptions.removable).toBe(true); // "removable" is not encoded in owc-json; falling back to "true"
-            expect(vlayerOptions.filtertype).toBe("Overlays"); // all data in owc-json will - for now - be an overlay. Might be changed in the future. 
-            expect(vlayerOptions.type).toBe("geojson"); // default
-            expect(vlayerOptions.url).toBe(operation.href.substr(0, operation.href.indexOf("?")));
-  
-            // @TODO: checke hier auch noch die vlayerOptions.params
+            let layertype = service.getLayertypeFromOfferingCode(offering);
+            if (layertype == WfsLayertype || layertype == GeojsonLayertype) {
+
+              const vlayerOptions = service.createVectorLayerFromOffering(offering, resource);
+    
+              expect(vlayerOptions.name).toBe(resource.properties.title);
+              expect(vlayerOptions.id as string).toBe(resource.id as string);
+              expect(vlayerOptions.opacity).toBe(1); // opacity is not encoded in owc-json per default. Allways falling back to 1.
+              // if active is not set in conext file, then it must default to true 
+              expect(vlayerOptions.visible).toBe(resource.properties.active === undefined ? true : resource.properties.active); 
+              expect(vlayerOptions.removable).toBe(true); // "removable" is not encoded in owc-json; falling back to "true"
+              expect(vlayerOptions.filtertype).toBe("Overlays"); // all data in owc-json will - for now - be an overlay. Might be changed in the future. 
+              expect(vlayerOptions.type).toBe("geojson"); // default
+              expect(vlayerOptions.url).toBe(operation.href.substr(0, operation.href.indexOf("?")));
+    
+              // @TODO: checke hier auch noch die vlayerOptions.params
+            }
           }
         }
       }
@@ -126,10 +147,12 @@ describe('OwcJsonService', () => {
   it('#getLegendUrl should return a proper url', () => {
     const service: OwcJsonService = TestBed.get(OwcJsonService);
     for(const context of allTestContexts) {
+      console.log("======= context: ", context.id)
       for(const resource of service.getResources(context)) {
         for(const offering of resource.properties.offerings) {
-
+          console.log("====== offering: ", offering);
           let legendUrl = service.getLegendUrl(offering);
+          console.log("====== legendUrl: ", legendUrl);
           let legendUrlS: string = legendUrl as string;
           expect(legendUrlS.lastIndexOf("//")).toBeTruthy();
 
@@ -138,4 +161,93 @@ describe('OwcJsonService', () => {
     }
   });
 
+
+});
+
+describe('OwcJsonService: writing data into owc', () => {
+  const allTestContexts = [barebonesContext, basicContext, exampleContext, coastalXTestContext];
+  let injector: TestBed;
+  let service: OwcJsonService;
+  
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [],
+      providers: [OwcJsonService, LayersService]
+    });
+    injector = getTestBed();
+    service = injector.get(OwcJsonService);
+  });
+
+  afterEach(() => {
+  });
+
+  it('#getLayers should properly restore a selection of layers from owc format created with #generateOwsContextFrom', () => {
+    const service: OwcJsonService = TestBed.get(OwcJsonService);
+    const layersService: LayersService = TestBed.get(LayersService);
+    const osm_layer = new osm(<any>{
+      visible: true,
+      legendImg: null
+    });
+    layersService.addLayer(osm_layer, 'Baselayers');
+    layersService.getBaseLayers().subscribe(baselayers => {
+      const owc = service.generateOwsContextFrom("someid", baselayers, [], [-190, -90, 190, 90]);
+      const layers = service.getLayers(owc);
+      expect(layers.length).toBeTruthy();
+    })
+  });
+
+  it('#generateOwcContextFrom should properly store and restore a geojson-layer', () => {
+
+    // creating test-layer
+    interface Props {
+      "name": string,
+      "id": string
+    };
+    let features: Feature<Polygon, Props>[] = [{
+      type: "Feature",
+      properties: {
+        name: "Testfeature",
+        id: "testfeatureId"
+      },
+      geometry: {
+        type: "Polygon",
+        coordinates: [[[0, 1], [0, 3]], [[2, 3], [2, 1]]]
+      }
+    }];
+    let featureCollection: FeatureCollection<Polygon, Props> = {
+      type: "FeatureCollection",
+      features: features
+    }
+
+    let options = {
+      style: new Style({
+        stroke: new Stroke({
+          color: 'rgb(214, 102, 27)',
+          width: 3
+        }),
+        fill: new Fill({
+          color: 'rgba(226, 123, 54, 0.1)'
+        })
+      })
+    };
+
+    let geojsonLayer = new VectorLayer({
+      name: "GeojsonLayer",
+      id: "GeojsonLayer",
+      type: GeojsonLayertype,
+      data: featureCollection,
+      options: options
+    });
+
+    // enconding and deconding
+    let context = service.generateOwsContextFrom("testcontext", [], [geojsonLayer], [-190, -90, 190, 90]);
+    let recoveredLayers = service.getLayers(context);
+    let recoveredLayer = recoveredLayers[0] as VectorLayer;
+
+    // testing
+    expect(recoveredLayer.id).toEqual(geojsonLayer.id);
+    expect(JSON.parse(recoveredLayer.data)).toEqual(geojsonLayer.data);
+    // expect(recoveredLayer.options).toEqual(geojsonLayer.options); // we dont encode style.
+
+  });
 });