+++ projects/services-owc-json/src/lib/owc-json.service.ts
@@ -1,8 +1,49 @@
 
 import { Injectable } from '@angular/core';
-import { IOwsContext, IOwsResource, IOwsOffering, IOwsOperation, IOwsContent, IEocOwsContext, IEocOwsResource, IEocOwsOffering } from '@ukis/datatypes-owc-json';
-import { ILayerGroupOptions, ILayerOptions, IRasterLayerOptions, VectorLayer, RasterLayer, IVectorLayerOptions, Layer } from '@ukis/datatypes-layers';
+import { IOwsContext, IOwsResource, IOwsOffering, IOwsOperation, IOwsContent, IEocOwsContext, IEocOwsResource, IEocOwsOffering, WMS_Offering, WFS_Offering, WCS_Offering, CSW_Offering, WMTS_Offering, GML_Offering, KML_Offering, GeoTIFF_Offering, GMLJP2_Offering, GMLCOV_Offering, GeoJson_Offering, Xyz_Offering } from '@ukis/datatypes-owc-json';
+import { ILayerGroupOptions, ILayerOptions, IRasterLayerOptions, VectorLayer, RasterLayer, IVectorLayerOptions, Layer, TLayertype, WmsLayertype, WmtsLayertype, WfsLayertype, GeojsonLayertype, CustomLayer, CustomLayertype, XyzLayertype, TRasterLayertype, isRasterLayertype, isVectorLayertype, TVectorLayertype } from '@ukis/datatypes-layers';
 import { TGeoExtent } from '@ukis/datatypes-map-state';
+import { ReplaceSource } from 'webpack-sources';
+
+
+
+
+export function isWmsOffering(str: string): str is WMS_Offering {
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/wms' || str == 'http://schemas.opengis.net/wms/1.1.1' || str == 'http://schemas.opengis.net/wms/1.1.0';
+}
+export function isWfsOffering(str: string): str is WFS_Offering {
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/wfs';
+}
+export function isWpsOffering(str: string): str is WCS_Offering {
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/wcs';
+}
+export function isCswOffering(str: string): str is CSW_Offering {
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/csw';
+}
+export function isWmtsOffering(str: string): str is WMTS_Offering {
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/wmts';
+}
+export function isGmlOffering(str: string): str is GML_Offering {
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/gml';
+}
+export function isKmlOffering(str:string): str is KML_Offering { 
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/kml';
+}
+export function isGeoTIFFOffering(str:string): str is GeoTIFF_Offering { 
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/geotiff';
+}
+export function isGMLJP2Offering(str:string): str is GMLJP2_Offering { 
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/gmljp2';
+}
+export function isGMLCOVOffering(str:string): str is GMLCOV_Offering { 
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/gmlcov';
+}
+export function isXyzOffering(str: string): str is Xyz_Offering {
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/xyz';
+}
+export function isGeoJsonOffering (str: string): str is GeoJson_Offering {
+  return str == 'http://www.opengis.net/spec/owc-geojson/1.0/req/geojson';
+}
 
 /**
  * OWS Context Service 
@@ -88,49 +129,6 @@ export class OwcJsonService {
     return active;
   }
 
-  //https://github.com/Terradue/ows-context-demo/blob/master/src/main/demo/index.html#L245:2
-  //create ukis layers ----------------------------------------------------------------------
-  layerGroupFromResource(resource: IOwsResource): ILayerGroupOptions {
-
-    throw new Error("This method has not been implemented yet.");
-
-    let layerGroupOptions: ILayerGroupOptions = {
-      filtertype: 'Overlays',
-      id: resource.id.toString(),
-      name: resource.properties.title,
-      removable: true,
-      layerRemovable: false,
-      layers: []
-    };
-
-    if (resource.properties.bbox) {
-      layerGroupOptions.bbox = resource.properties.bbox;
-    }
-
-    return layerGroupOptions;
-  }
-
-  getlayersFromResource(resource: IOwsResource): ILayerOptions {
-
-    throw new Error("This method has not been implemented yet.");
-
-    let layers: ILayerOptions[] = [];
-    let offerings = this.getResourceOfferings(resource);
-
-    let layerOptions: ILayerOptions = {
-      name: resource.properties.title,
-      id: resource.id.toString(),
-      visible: true,
-      type: 'custom',
-      removable: true,
-      attribution: '',
-      continuousWorld: false,
-      opacity: 1
-    }
-
-    return layerOptions;
-  }
-
   getResourceOpacity(resource: IOwsResource): number {
     let opacity = 1;
     if (resource.properties.hasOwnProperty("opacity")) {
@@ -155,12 +153,13 @@ export class OwcJsonService {
 
 
   /** Offering --------------------------------------------------- */
-  getOfferingCode(offering: IOwsOffering): "wms" | "wmts" | "wfs" {
-    for (let part of offering.code.split('/')) {
-      if (["wms", "wfs", "wmts"].includes(part.toLowerCase())) {
-        return part as "wms" | "wmts" | "wfs";
-      }
-    }
+  getLayertypeFromOfferingCode(offering: IOwsOffering): TLayertype {
+    if(isWmsOffering(offering.code)) return WmsLayertype;
+    if(isWmtsOffering(offering.code)) return WmtsLayertype;
+    if(isWfsOffering(offering.code)) return WfsLayertype;
+    if(isGeoJsonOffering(offering.code)) return GeojsonLayertype;
+    if(isXyzOffering(offering.code)) return XyzLayertype;
+    else return offering.code; // an offering can also be any other string. 
   }
 
   checkIfServiceOffering(offering: IOwsOffering) {
@@ -189,7 +188,6 @@ export class OwcJsonService {
     if (offering.hasOwnProperty("styles")) {
       let defaultStyle = offering.styles.filter(style => style.default);
       if (defaultStyle.length > 0) {
-        console.log(defaultStyle[0].legendURL);
         return defaultStyle[0].legendURL;
       }
     } else if (offering.hasOwnProperty("legendUrl")) {
@@ -212,9 +210,26 @@ export class OwcJsonService {
 
 
   createVectorLayerFromOffering(offering: IOwsOffering, resource: IOwsResource, context?: IOwsContext): VectorLayer {
+    let layerType = this.getLayertypeFromOfferingCode(offering);
+    
+    if(!isVectorLayertype(layerType)){
+      console.error(`This type of layer '${layerType}' / offering '${offering.code}' cannot be converted into a Vectorlayer`); 
+      return null; 
+    }
+
     let iconUrl = this.getIconUrl(offering);
-    let layerUrl = this.getUrlFromUri(offering.operations[0].href);
-    let params = this.getJsonFromUri(offering.operations[0].href);
+    
+    let layerUrl, params;
+    // if we have a operations-offering (vs. a data-offering):
+    if(offering.operations) layerUrl = this.getUrlFromUri(offering.operations[0].href);
+    if(offering.operations) params = this.getJsonFromUri(offering.operations[0].href);
+    
+    let data;
+    // if we have a data-offering: 
+    if(offering.contents) {
+      data = offering.contents[0].content;
+    }
+
     let legendUrl = this.getLegendUrl(offering);
 
     let layerOptions: IVectorLayerOptions = {
@@ -222,13 +237,14 @@ export class OwcJsonService {
       name: this.getResourceTitle(resource),
       displayName: this.getDisplayName(offering, resource),
       visible: this.isActive(resource),
-      type: 'geojson',
+      type: layerType,
       removable: true,
       attribution: this.getResourceAttribution(resource),
       continuousWorld: false,
       opacity: this.getResourceOpacity(resource),
       url: layerUrl ? layerUrl : null,
-      legendImg: legendUrl ? legendUrl : null
+      legendImg: legendUrl ? legendUrl : null,
+      data: data
     };
 
 
@@ -247,26 +263,30 @@ export class OwcJsonService {
 
 
   createRasterLayerFromOffering(offering: IOwsOffering, resource: IOwsResource, context?: IOwsContext): RasterLayer {
-    let offeringCode = this.getOfferingCode(offering);
+    let layerType = this.getLayertypeFromOfferingCode(offering);
+    
+    if(!isRasterLayertype(layerType)) {
+      console.error(`This type of offering '${offering.code}' cannot be converted into a rasterlayer.`);
+      return null;
+    }
 
     let customParams;
-    switch (offeringCode as "wms" | "wmts" | "xyz" | "geojson" | "custom") {
-      case "xyz":
-      case "wms":
-      case "geojson":
-      case "custom":
+    switch (layerType) {
+      case WmsLayertype:
+      case CustomLayertype:
         customParams = this.getWmsSpecificParamsFromOffering(offering, resource);
         break;
-      case "wmts":
+      case WmtsLayertype:
         customParams = this.getWmtsSpecifiParamsFromOffering(offering, resource);
         break;
-      default:
-        throw new Error("A service of type ${offeringCode} cannot be converted to a RasterLayer object.");
+      case XyzLayertype:
+        // xyz and wts are simple, pure rest services with no "operations" . 
+        break;
     }
 
     let layerOptions: IRasterLayerOptions = {
       id: resource.id as string,
-      type: offeringCode as "wms" | "wmts" | "xyz" | "geojson" | "custom",
+      type: layerType,
       removable: true,
       continuousWorld: false,
       opacity: this.getResourceOpacity(resource),
@@ -326,6 +346,9 @@ export class OwcJsonService {
     return params;
   }
 
+  /**
+   * @TODO: at the moment this function returns a layer for the first offering of every feature. Should it maybe chose  one of the offerings by type?
+   */
   public getLayers(owc: IOwsContext): Layer[] {
     let features = owc.features;
     let layers = []
@@ -333,12 +356,17 @@ export class OwcJsonService {
     /** layer priority first wms then wms then wfs if multiple in the offerings */
     for (let feature of features) {
       let offerings = feature.properties.offerings;
+      let indx = 0;
       for (let offering of offerings) {
-        let layer;
-        if (offering.code.toLocaleLowerCase().indexOf('wmts') != -1 || offering.code.toLocaleLowerCase().indexOf('wms') != -1) {
-          layers.push(this.createRasterLayerFromOffering(offering, feature, owc))
-        } else if (offering.code.toLocaleLowerCase().indexOf('wfs') != -1 || offering.code.toLocaleLowerCase().indexOf('geojson') != -1) {
-          layers.push(this.createVectorLayerFromOffering(offering, feature, owc))
+        if (indx > 0) break;
+        indx += 1;
+        let layerType = this.getLayertypeFromOfferingCode(offering);
+        if(isRasterLayertype(layerType)) {
+          layers.push(this.createRasterLayerFromOffering(offering, feature, owc));
+        } else if(isVectorLayertype(layerType)) {
+          layers.push(this.createVectorLayerFromOffering(offering, feature, owc));
+        } else {
+          console.error(`This type of service (${layerType}) has not been implemented yet.`);
         }
       }
     }
@@ -389,15 +417,23 @@ export class OwcJsonService {
 
   /**
    * @TODO:
-   *   - bounding box
    *   - properties
    */
-  generateOwsContextFrom(id: string, baselayers: Layer[], overlays: Layer[], extent?: TGeoExtent): IEocOwsContext {
+  generateOwsContextFrom(id: string, baselayers: Layer[], overlays: Layer[], extent?: TGeoExtent, properties?): IEocOwsContext {
+
+    if(!properties) {
+      properties = {
+        lang: "",
+        links: [], 
+        title: "", 
+        updated: ""
+      };
+    }
 
     let owc: IEocOwsContext = {
       "id": id,
       "type": "FeatureCollection",
-      "properties": null,
+      "properties": properties,
       "features": []
     };
 
@@ -418,7 +454,6 @@ export class OwcJsonService {
     return owc;
   }
 
-
   generateResourceFromLayer(layer: Layer): IEocOwsResource {
     let resource: IEocOwsResource = {
       "id": layer.id,
@@ -427,7 +462,7 @@ export class OwcJsonService {
         updated: null,
         offerings: [this.generateOfferingFromLayer(layer)],
         opacity: layer.opacity,
-        attribution: layer.attribution
+        attribution: layer.attribution,
       },
       "type": "Feature",
       "geometry": null
@@ -438,41 +473,65 @@ export class OwcJsonService {
   generateOfferingFromLayer(layer: Layer, legendUrl?: string, iconUrl?: string): IEocOwsOffering {
     let offering: IEocOwsOffering = {
       "code": this.getOfferingCodeFromLayer(layer),
-      "operations": this.getOperationsFromLayer(layer),
       "title": layer.name
     };
+
+    if(layer.type == GeojsonLayertype) {
+      offering.contents = this.getContentsFromLayer(layer as VectorLayer);
+    } else {
+      offering.operations = this.getOperationsFromLayer(layer);
+    }
+
     if(legendUrl) offering.legendUrl = legendUrl;
     if(iconUrl) offering.iconUrl = iconUrl;
+    
     return offering;
   }
 
   getOfferingCodeFromLayer(layer: Layer): string {
-
     switch (layer.type) {
-      case "wms":
+      case WmsLayertype:
         return 'http://www.opengis.net/spec/owc-geojson/1.0/req/wms';
-      case "wmts":
+      case WmtsLayertype:
         return 'http://www.opengis.net/spec/owc-geojson/1.0/req/wmts';
-      case "geojson":
+      case GeojsonLayertype:
         return 'http://www.opengis.net/spec/owc-geojson/1.0/req/geojson';
-      case "xyz":
+      case XyzLayertype:
         return 'http://www.opengis.net/spec/owc-geojson/1.0/req/xyz';
       default:
-        throw new Error("This type of service (" + layer.type + ") has not been implemented yet.");
+        console.error(`This type of layer (${layer.type}) has not been implemented yet.`);
+        return null;
     }
+  }
+
+  getContentsFromLayer(layer: VectorLayer): IOwsContent[] {
+    let contents = [];
+    switch(layer.type) {
+      case GeojsonLayertype:
+        let content = {
+          type: "FeatureCollection",
+          content: JSON.stringify(layer.data)
+        };
+        contents.push(content);
+      break;
+      default: 
+        console.error(`Cannot get contents for this type of vectorlayer: (${layer.type})`);
+    }
+    return contents;
   }
 
   getOperationsFromLayer(layer: Layer): IOwsOperation[] {
     if (layer instanceof RasterLayer) {
       switch (layer.type) {
-        case "wms":
+        case WmsLayertype:
           return this.getWmsOperationsFromLayer(layer);
-        case "wmts":
+        case WmtsLayertype:
           return this.getWmtsOperationsFromLayer(layer);
-        case "xyz":
-          return this.getTmsOperationsFromLayer(layer);
+        case XyzLayertype:
+          return this.getXyzOperationsFromLayer(layer);
         default:
-          throw new Error("This type of service (" + layer.type + ") has not been implemented yet.");
+          console.error(`Cannot get operations for this type of layer: (${layer.type})`);
+          return [];
       }
     }
 
@@ -480,20 +539,29 @@ export class OwcJsonService {
       switch (layer.type) {
         // case "wfs": <--- this type of layer has not been implemented yet in datatypes-layers/Layers.ts 
         //   return this.getWfsOperationsFromLayer(layer);
-        case "geojson":
-          return this.getGeojsonOperationsFromLayer(layer);
         default:
-          throw new Error("This type of service (" + layer.type + ") has not been implemented yet.");
+          console.error(`This type of service (${layer.type}) has not been implemented yet.`);
+          return [];
       }
     }
 
   }
 
-  getGeojsonOperationsFromLayer(layer: VectorLayer): IOwsOperation[] {
-    // A layer with layer.type == "geojson" is merely a *.geojson file; thus no operations are defined. 
-    return [];
-  }
 
+  getXyzOperationsFromLayer(layer: RasterLayer): IOwsOperation[] {
+    let restCall: IOwsOperation = {
+      "code": "REST",
+      "method": "GET",
+      "type": "text/html",
+      "href": `${layer.url}`
+    }
+
+    let operations: IOwsOperation[] = [
+      restCall
+    ];
+
+    return operations;
+  }
 
   getTmsOperationsFromLayer(layer: RasterLayer): IOwsOperation[] {
     // @TODO: what operations are defined on TMS? https://wiki.osgeo.org/wiki/Tile_Map_Service_Specification
@@ -549,12 +617,15 @@ export class OwcJsonService {
     let url = layer.url;
     let wmsVersion = layer.params.VERSION;
     let layerName = layer.name;
+    let layerId = layer.id;
+    let format = "image/vnd.jpeg-png";
+    if (layer.params && layer.params.FORMAT) format = layer.params.FORMAT;
 
     let getMap: IOwsOperation = {
       "code": "GetMap",
       "method": "GET",
-      "type": "image/vnd.jpeg-png",
-      "href": `${url}?service=WMS&version=${wmsVersion}&request=GetMap&TRANSPARENT=TRUE&LAYERS=${layerName}&FORMAT=image/vnd.jpeg-png&TILED=true`
+      "type": format,
+      "href": `${url}?service=WMS&version=${wmsVersion}&request=GetMap&TRANSPARENT=TRUE&LAYERS=${layerId}&FORMAT=${format}&TILED=true`
     };
 
     let getCapabilities: IOwsOperation = {
@@ -568,7 +639,7 @@ export class OwcJsonService {
       "code": "GetFeatureInfo",
       "method": "GET",
       "type": "text/html",
-      "href": `${url}?service=WMS&version=${wmsVersion}&request=GetFeatureInfo&TRANSPARENT=TRUE&LAYERS=${layerName}&FORMAT=image/vnd.jpeg-png`
+      "href": `${url}?service=WMS&version=${wmsVersion}&request=GetFeatureInfo&TRANSPARENT=TRUE&LAYERS=${layerId}&FORMAT=${format}`
     }
 
     let operations: IOwsOperation[] = [
@@ -585,12 +656,15 @@ export class OwcJsonService {
     let url = layer.url;
     let wmtsVersion = layer.params.version;
     let layerName = layer.name;
+    let layerId = layer.id;
+    let format = "image/vnd.jpeg-png";
+    if (layer.params && layer.params.FORMAT) format = layer.params.FORMAT;
 
     let getTile: IOwsOperation = {
       "code": "GetTile",
-      "href": `${url}?SERVICE=WMTS&REQUEST=GetTile&FORMAT=image%2Fpng&LAYER=${layerName}&VERSION=${wmtsVersion}`,
+      "href": `${url}?SERVICE=WMTS&REQUEST=GetTile&FORMAT=${format}&LAYER=${layerId}&VERSION=${wmtsVersion}`,
       "method": "GET",
-      "type": "image/png"
+      "type": format
     };
 
     let getCapabilities: IOwsOperation = {