+++ projects/user-info/src/lib/login/login.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { LoginComponent } from './login.component';
-import { UserService } from '@ukisuser-info/src/lib/user.service';
+import { UserService } from '@ukis/user-info/src/lib/user.service';
 /*
 describe('LoginComponent', () => {
   let component: LoginComponent;